Decompose the ToggleMergeUnmergeButton into its own class (or component)
- Removed FieldNameCellFactory.java and MergeableFieldCell.java
- Made it easier to add more side buttons in the future
